2D Rutherford Scattering
Models the famous Rutherford scattering in 2D. Prompts user to give the initial velocity when shooting the alpha particle. Also prompts user to enter the impact parameter. Returns the scattering angle and tells the user whether the two particles touch. If the particles do not touch, returns the minimum distance between the surfaces of the two particles during the collision.